Move GridInfo to dedicated module
